d.o #2281491, Refactoring on the suggestion of Dave Reid
